requirements.txt
streamlit
pandas
supabase
ccxt
plotly
google-generativeai
google-search-results
python-dotenv
schedule
feedparser
requests
beautifulsoup4